requirements.txt
google-cloud-bigquery==3.11.4
plost
matplotlib
snowflake-snowpark-python[pandas]
streamlit
plotly